requirements.txt
faiss-cpu==1.11.0
openai==1.82.0
python-dotenv==1.1.0
tiktoken==0.9.0
